# openapi.yaml
openapi: 3.0.3
info:
  title: Authentication API
  description: API for user registration, login, OTP, and session management.
  version: 1.0.0

servers:
  - url: http://localhost:3000/api
    description: Route Root

tags:
  - name: Authentication
    description: All authentication-related endpoints

paths:
  /sign-up:
    post:
      tags: [Authentication]
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                device_name:
                  type: string
                user_agent:
                  type: string
                os:
                  type: string
                ip:
                  type: string
                type:
                  type: string
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              example:
                success: true
                message: User Created Successfully!
                data:
                  refresh_token: "..."
                  access_token: "..."
                  user: {}
        '403':
          description: Missing email or password
        '406':
          description: User already exists

  /sign-in:
    post:
      tags: [Authentication]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '404':
          description: User not found
        '405':
          description: Invalid password
        '420':
          description: Device not trusted
        '421':
          description: No active session

  /send-otp:
    post:
      tags: [Authentication]
      summary: Send OTP for verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, user_id]
              properties:
                email:
                  type: string
                user_id:
                  type: string
      responses:
        '200':
          description: OTP sent successfully
        '403':
          description: Missing data
        '500':
          description: Failed to send OTP

  /verify-otp:
    post:
      tags: [Authentication]
      summary: Verify the sent OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, user_id, otp]
              properties:
                email:
                  type: string
                user_id:
                  type: string
                otp:
                  type: string
      responses:
        '200':
          description: OTP verified successfully
        '403':
          description: Missing data
        '500':
          description: OTP expired or verification failed

  /session-device:
    post:
      tags: [Authentication]
      summary: Create or update session for a specific device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, device_name, os, user_id, ip]
              properties:
                email:
                  type: string
                device_name:
                  type: string
                user_agent:
                  type: string
                os:
                  type: string
                user_id:
                  type: string
                ip:
                  type: string
                type:
                  type: string
      responses:
        '200':
          description: Device session set or updated successfully
        '403':
          description: Missing required data

  /check:
    post:
      tags: [Authentication]
      summary: Validate a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Refresh token is valid
        '500':
          description: Refresh token is invalid or expired